タイマーやダイアログをサービスとして切り出し依存関係を整理。

+++ VoiceOfClock/UseCases/PeriodicTimerLifetimeManager.cs
@@ -1,5 +1,5 @@
-﻿using CommunityToolkit.Mvvm.Messaging;
-using CommunityToolkit.Mvvm.Messaging.Messages;
+﻿using CommunityToolkit.Diagnostics;
+using CommunityToolkit.Mvvm.Messaging;
 using CommunityToolkit.WinUI.Helpers;
 using I18NPortable;
 using Microsoft.UI.Dispatching;
@@ -13,48 +13,73 @@ using System.Reactive.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using VoiceOfClock.Contract.Services;
-using VoiceOfClock.Contract.UseCases;
-using VoiceOfClock.Models.Domain;
+using VoiceOfClock.Contracts.Services;
+using VoiceOfClock.Contracts.UseCases;
+using VoiceOfClock.Core.Contracts.Services;
+using VoiceOfClock.Core.Domain;
 
 namespace VoiceOfClock.UseCases;
 
-public sealed class PeriodicTimerLifetimeManager : IApplicationLifeCycleAware
+
+public sealed class PeriodicTimerLifetimeManager 
+    : IApplicationLifeCycleAware
     , IRecipient<ActiveTimerCollectionRequestMessage>
 {
     private readonly IMessenger _messenger;
     private readonly ISoundContentPlayerService _soundContentPlayerService;
+    private readonly ITimeTriggerService _timeTriggerService;
     private readonly PeriodicTimerRepository _periodicTimerRepository;
-    private readonly ObservableCollection<PeriodicTimerRunningInfo> _timers;
-    private readonly DispatcherQueueTimer _timer;
 
-    public ReadOnlyObservableCollection<PeriodicTimerRunningInfo> Timers { get; }
-    public PeriodicTimerRunningInfo InstantPeriodicTimer { get; }
+    private const string TimeTriggerGroupId = nameof(PeriodicTimerLifetimeManager);
 
-    public PeriodicTimerLifetimeManager(IMessenger messenger,
+    public PeriodicTimerLifetimeManager(
+        IMessenger messenger,
         ISoundContentPlayerService soundContentPlayerService,
+        ITimeTriggerService timeTriggerService,
         PeriodicTimerRepository periodicTimerRepository
         )
     {
         _messenger = messenger;
         _soundContentPlayerService = soundContentPlayerService;
+        _timeTriggerService = timeTriggerService;
         _periodicTimerRepository = periodicTimerRepository;
-        _timers = new ObservableCollection<PeriodicTimerRunningInfo>(_periodicTimerRepository.ReadAllItems().OrderBy(x => x.Order).Select(x => new PeriodicTimerRunningInfo(x, _periodicTimerRepository)));
-        Timers = new(_timers);
-        _timer = DispatcherQueue.GetForCurrentThread().CreateTimer();
-        _timer.Interval = TimeSpan.FromSeconds(0.2);
-        _timer.IsRepeating = true;
-        _timer.Tick += _timer_Tick;
-        _timer.Start();
 
-        InstantPeriodicTimer = new PeriodicTimerRunningInfo(new PeriodicTimerEntity() 
-        {
-            IsEnabled = false,
-            Id = Guid.Empty,
-            Title = "InstantPeriodicTimer_Title".Translate(),                
-        }, _periodicTimerRepository);
+        _timeTriggerService.TimeTriggered += _timeTriggerService_TimeTriggered;
     }
 
+    private void _timeTriggerService_TimeTriggered(object? sender, TimeTriggeredEventArgs e)
+    {
+        if (e.GroupId != TimeTriggerGroupId) { return; }
+
+        if (Guid.TryParse(e.Id, out Guid timerId) is false) { return; }
+
+        var timer = timerId == InstantPeriodicTimerId ? InstantPeriodicTimer : _periodicTimerRepository.FindById(timerId);
+        
+        Guard.IsNotNull(timer);
+
+        if (TimerIsInsidePeriod(timer) is false)
+        {
+            _ = SendCurrentTimeVoiceAsync(e.TriggerTime);
+            Debug.WriteLine($"ピリオドダイマー： {timer.Title} が完了");
+
+            // 繰り返しが無い場合は無効に切り替える
+            if (timer.EnabledDayOfWeeks.Any() is false)
+            {
+                timer.IsEnabled = false;
+                _periodicTimerRepository.UpdateItem(timer);
+                _messenger.Send(new UpdatePeriodicTimerMessage(timer));
+            }
+        }
+        else
+        {
+            // 次に通知すべき時間を割り出す
+            _ = SendCurrentTimeVoiceAsync(e.TriggerTime);
+            Debug.WriteLine($"ピリオドダイマー： {timer.Title} の再生を開始");
+
+            _timeTriggerService.SetTimeTrigger(e.Id.ToString(), e.TriggerTime + timer.IntervalTime, TimeTriggerGroupId);
+            _messenger.Send(new ProgressPeriodPeriodicTimerMessage(timer));
+        }
+    }
 
     void IApplicationLifeCycleAware.Initialize()
     {
@@ -64,7 +89,11 @@ public sealed class PeriodicTimerLifetimeManager : IApplicationLifeCycleAware
         {
             CreatePeriodicTimer("PeriodicTimer_TemporaryTitle".Translate(1), TimeSpan.FromHours(9), TimeSpan.FromHours(10), TimeSpan.FromMinutes(5), Enum.GetValues<DayOfWeek>(), isEnabled: false);
         }
+
+        _timeTriggerService.SetTimeTriggerGroup(TimeTriggerGroupId, GetTimers().Select(x => (x.Id.ToString(), GetNextTime(x))));
     }
+
+
 
     void IApplicationLifeCycleAware.Resuming() { }
 
@@ -78,46 +107,70 @@ public sealed class PeriodicTimerLifetimeManager : IApplicationLifeCycleAware
         }
     }
 
-
-    DateTime _lastTickTime = DateTime.MinValue;
-    private void _timer_Tick(DispatcherQueueTimer sender, object args)
+    IEnumerable<PeriodicTimerEntity> GetInsideTimers()
     {
-        if (GetInsideEnablingTimeTimers() is not null and var timers && timers.Any())
+        return _periodicTimerRepository.ReadAllItems().Where(TimerIsInsidePeriod);
+    }
+
+    public IEnumerable<PeriodicTimerEntity> GetTimers()
+    {
+        return _periodicTimerRepository.ReadAllItems();
+    }
+
+
+    public static DateTime GetNextTime(PeriodicTimerEntity entity)
+    {
+        return GetNextTime(entity, DateTime.Now);
+    }
+
+    public static DateTime GetNextTime(PeriodicTimerEntity entity, DateTime targetTime)
+    {
+        if (TimerIsInsidePeriod(entity, targetTime))
         {
-            foreach (var timer in timers)
-            {
-                // 次に通知すべき時間を割り出す
-                if (timer.NextTime < DateTime.Now)
-                {
-                    if (timer.NextTime.TimeOfDay == timer.StartTime)
-                    {
-                        _ = SendCurrentTimeVoiceAsync(timer.NextTime);
-                        Debug.WriteLine($"ピリオドダイマー： {timer.Title} を開始");
-                        
-                        timer.IncrementNextTime();
-                    }
-                    else if (timer.NextTime.TimeOfDay == timer.EndTime)
-                    {
-                        _ = SendCurrentTimeVoiceAsync(timer.NextTime);
-                        Debug.WriteLine($"ピリオドダイマー： {timer.Title} が完了");
-                        
-                        timer.OnEnded();
-                    }
-                    else
-                    {
-                        _ = SendCurrentTimeVoiceAsync(timer.NextTime);
-                        Debug.WriteLine($"ピリオドダイマー： {timer.Title} の再生を開始");
+            var (_, _, nextTime) = InsidePeriodCulcNextTime(entity, targetTime);
+            return nextTime;
+        }
+        else
+        {
+            return OutsideCulcNextTime(entity);
+        }
+    }
 
-                        timer.IncrementNextTime();
-                    }
-                }
+    public static bool TimerIsInsidePeriod(PeriodicTimerEntity entity)
+    {
+        return TimeHelpers.IsInsideTime(DateTime.Now.TimeOfDay, entity.StartTime, entity.EndTime);
+    }
 
-                timer.UpdateElapsedTime();
-            }
+    public static bool TimerIsInsidePeriod(PeriodicTimerEntity entity, DateTime targetTime)
+    {        
+        return TimeHelpers.IsInsideTime(targetTime.TimeOfDay, entity.StartTime, entity.EndTime);
+    }
+
+
+    public static (DateTime startDateTime, TimeSpan elapsedTime, DateTime nextTime) InsidePeriodCulcNextTime(PeriodicTimerEntity entity)
+    {
+        return InsidePeriodCulcNextTime(entity, DateTime.Now);
+    }
+
+    public static (DateTime startDateTime, TimeSpan elapsedTime, DateTime nextTime) InsidePeriodCulcNextTime(PeriodicTimerEntity entity, DateTime targetTime)
+    {
+        DateTime now = targetTime;        
+        var startDateTime = now.Date + entity.StartTime;
+        if (entity.StartTime > now.TimeOfDay)
+        {
+            startDateTime -= TimeSpan.FromDays(1);
         }
 
-        _lastTickTime = DateTime.Now;
-        //Debug.WriteLine(_lastTickTime.TimeOfDay);
+        var elapsedTime = (now - startDateTime).TrimMilliSeconds();
+        int count = (int)Math.Ceiling(elapsedTime / entity.IntervalTime);
+        var nextTime = now.Date + entity.StartTime + entity.IntervalTime * count;
+
+        return (startDateTime, elapsedTime, nextTime);
+    }
+
+    public static DateTime OutsideCulcNextTime(PeriodicTimerEntity entity, DateTime? targetTime = null)
+    {
+        return TimeHelpers.CulcNextTime(targetTime ?? DateTime.Now, entity.StartTime, entity.EnabledDayOfWeeks);
     }
 
     Task SendCurrentTimeVoiceAsync(DateTime time)
@@ -125,7 +178,17 @@ public sealed class PeriodicTimerLifetimeManager : IApplicationLifeCycleAware
         return _soundContentPlayerService.PlayTimeOfDayAsync(time);
     }
 
-    IEnumerable<PeriodicTimerRunningInfo> GetInsideEnablingTimeTimers()
+    public PeriodicTimerEntity InstantPeriodicTimer { get; } = new PeriodicTimerEntity() 
+    {
+        Id = InstantPeriodicTimerId,
+        IsEnabled = false,
+        SoundSourceType = SoundSourceType.DateTimeToSpeech,        
+        SoundContent = string.Empty,
+        StartTime = TimeSpan.Zero,
+        EndTime = TimeSpan.FromDays(1) - TimeSpan.FromSeconds(1),
+    };
+
+    IEnumerable<PeriodicTimerEntity> GetInsideEnablingTimeTimers()
     {
         if (InstantPeriodicTimer.IsEnabled)
         {
@@ -133,13 +196,13 @@ public sealed class PeriodicTimerLifetimeManager : IApplicationLifeCycleAware
         }
 
         TimeSpan timeOfDay = DateTime.Now.TimeOfDay.TrimMilliSeconds();
-        foreach (var timer in _timers.Where(x => x.IsEnabled && TimeHelpers.IsInsideTime(timeOfDay, x.StartTime, x.EndTime)))
+        foreach (var timer in GetInsideTimers().Where(x => x.IsEnabled))
         {
             yield return timer;
         }
     }        
 
-    public PeriodicTimerRunningInfo CreatePeriodicTimer(string title, TimeSpan startTime, TimeSpan endTime, TimeSpan intervalTime, DayOfWeek[] enabledDayOfWeeks, bool isEnabled = true)
+    public PeriodicTimerEntity CreatePeriodicTimer(string title, TimeSpan startTime, TimeSpan endTime, TimeSpan intervalTime, DayOfWeek[] enabledDayOfWeeks, bool isEnabled = true)
     {
         var entity = _periodicTimerRepository.CreateItem(new PeriodicTimerEntity 
         {
@@ -152,44 +215,72 @@ public sealed class PeriodicTimerLifetimeManager : IApplicationLifeCycleAware
             Order = int.MaxValue,
         });
 
-        var runningTimerInfo = new PeriodicTimerRunningInfo(entity, _periodicTimerRepository);
-        _timers.Add(runningTimerInfo);
-
         // 並び順を確実に指定する
-        foreach (var (timer, index) in _timers.Select((x, i) => (x, i)))
+        foreach (var (timer, index) in _periodicTimerRepository.ReadAllItems().OrderBy(x => x.Order).Select((x, i) => (x, i)))
         {
-            timer._entity.Order = index;
-            _periodicTimerRepository.UpdateItem(timer._entity);
+            timer.Order = index;
+            _periodicTimerRepository.UpdateItem(timer);
         }
 
-        return runningTimerInfo;
+        if (entity.IsEnabled)
+        {
+            _timeTriggerService.SetTimeTrigger(entity.Id.ToString(), GetNextTime(entity), TimeTriggerGroupId);
+        }
+
+        return entity;
     }
 
-    public bool DeletePeriodicTimer(PeriodicTimerRunningInfo timer)
+    
+
+    public bool DeletePeriodicTimer(PeriodicTimerEntity entity)
     {
-        var entity = timer._entity;
+        if (IsInstantPeriodicTimer(entity)) { return false; }
+
         var deleted1 = _periodicTimerRepository.DeleteItem(entity.Id);
-        var deleted2 = _timers.Remove(timer);        
-        return deleted2;
+        _timeTriggerService.DeleteTimeTrigger(entity.Id.ToString(), TimeTriggerGroupId);
+        return deleted1;
     }
 
+    public void UpdatePeriodicTimer(PeriodicTimerEntity entity)
+    {
+        if (IsInstantPeriodicTimer(entity)) { return; }
+
+        _periodicTimerRepository.UpdateItem(entity);
+        _timeTriggerService.SetTimeTrigger(entity.Id.ToString(), GetNextTime(entity), TimeTriggerGroupId);
+    }
+
+    public int GetTimerCount()
+    {
+        return _periodicTimerRepository.Count();
+    }
+
+    private static readonly Guid InstantPeriodicTimerId = Guid.Parse("DB31A3F3-AEE1-4C66-8C6D-B0ECAE756524");
+
+    public bool IsInstantPeriodicTimer(PeriodicTimerEntity entity) => entity.Id == InstantPeriodicTimerId;
+
+    public async ValueTask<bool> GetNowInstantPeriodicTimerEnabled()
+    {
+        return await _timeTriggerService.GetTimeTrigger(InstantPeriodicTimerId.ToString()) != null;
+    }
 
     public void StartInstantPeriodicTimer(TimeSpan intervalTime)
     {
-        TimeSpan timeOfDay = DateTime.Now.TimeOfDay;
-        timeOfDay = new TimeSpan(timeOfDay.Hours, timeOfDay.Minutes, timeOfDay.Seconds);
-        using (InstantPeriodicTimer.DeferUpdate())
-        {
-            InstantPeriodicTimer.IsEnabled = true;
-            InstantPeriodicTimer.IntervalTime = intervalTime;
-            InstantPeriodicTimer.StartTime = timeOfDay;
-            InstantPeriodicTimer.EndTime = timeOfDay - TimeSpan.FromSeconds(1);
-        }
+        var now = DateTime.Now;
+        InstantPeriodicTimer.IntervalTime = intervalTime;
+        InstantPeriodicTimer.StartTime = now.TimeOfDay;
+        InstantPeriodicTimer.EndTime = (now - TimeSpan.FromSeconds(1)).TimeOfDay;
+        TimeSpan timeOfDay = now.TimeOfDay;        
+        _timeTriggerService.SetTimeTrigger(
+            InstantPeriodicTimerId.ToString(),
+            now.Date + timeOfDay.TrimMilliSeconds(), 
+            TimeTriggerGroupId
+            );
+        _messenger.Send(new UpdatePeriodicTimerMessage(InstantPeriodicTimer));
     }
 
     public void StopInstantPeriodicTimer()
     {
-        InstantPeriodicTimer.IsEnabled = false;
+        _timeTriggerService.DeleteTimeTrigger(InstantPeriodicTimerId.ToString(), TimeTriggerGroupId);
     }
 
 }
